@@ -1,10 +1,10 @@
-import React, { useState } from 'react'
-import Paper from '@material-ui/core/Paper'
-import Grid from '@material-ui/core/Grid'
-import { makeStyles } from '@material-ui/core/styles'
+import React from 'react'
+import Grid from '@mui/material/Grid'
+import Paper from '@mui/material/Paper'
 import TodoItem from './Action'
+import { createTheme } from '@mui/material/styles'
 
-const useStyles = makeStyles((theme) => ({
+const theme = createTheme({
 	todo: {
 		maxWidth: 400,
 		margin: 'auto',
@@ -25,7 +25,7 @@ const useStyles = makeStyles((theme) => ({
 		width: '100%',
 	},
 	form: {
-		padding: theme.spacing(2),
+		padding: 2,
 	},
 	list: {
 		listStyle: 'none',
@@ -33,10 +33,9 @@ const useStyles = makeStyles((theme) => ({
 		marginBottom: 0,
 		borderRadius: '0 0 4px 4px',
 	},
-}))
+})
 
 const Todo = ({ data }) => {
-	const classes = useStyles()
 	const initialState = Object.keys(data['events']).map((key) => {
 		return {
 			day: key,
@@ -50,19 +49,19 @@ const Todo = ({ data }) => {
 	return (
 		<Grid
 			container
-			className={classes.todo}
+			className={theme.todo}
 			justify="center"
 			direction="column"
 		>
 			<header>
-				<h1 className={classes.srOnly}> Todo App </h1>
+				<h1 className={theme.srOnly}> Todo App </h1>
 			</header>
-			<Paper className={classes.paper} elevation={3}>
-				<ul className={classes.list}>
+			<Paper className={theme.paper} elevation={3}>
+				<ul className={theme.list}>
 					{initialState.map((todo) => (
 						<div>
 							<h2 style={{marginTop: "10px"}}>{new Date(todo.day).toLocaleDateString()}</h2>
-							<ul className={classes.list}>
+							<ul className={theme.list}>
 								{todo.tasks.map((task) => (
 									<TodoItem key={task.id} todo={task} updateTodo={() => {}} disabled={true} />
 								))}
