@@ -1,20 +1,9 @@
 import React, { useEffect } from 'react'
 import ReactDOM from 'react-dom'
 import Head from 'next/head'
-import { createMuiTheme } from '@material-ui/core/styles'
-import { ThemeProvider } from '@material-ui/styles'
-import CssBaseline from '@material-ui/core/CssBaseline'
-
-const theme = createMuiTheme({
-	palette: {
-		background: {
-			default: '#EEE',
-		},
-		primary: {
-			main: '#673ab7',
-		},
-	},
-})
+import { ThemeProvider } from '@mui/material/styles'
+import CssBaseline from '@mui/material/CssBaseline';
+import theme from '../src/theme';
 
 const MyApp = ({ Component, pageProps }) => {
 	useEffect(() => {
